+++ lib/src/core/utils/event_log.dart
@@ -2,13 +2,14 @@ import '../../domain/entities/log_event_entity.dart';
 import '../../domain/services/i_app_tracking_service.dart';
 
 class EventLog {
-  EventLog({required this.analytics});
+  EventLog({this.firebaseAnalytics});
 
-  final IAppTrackingService analytics;
+  final IAppTrackingService? firebaseAnalytics;
 
   Future<void> createEvent({required LogEventEntity event}) async {
     await Future.wait([
-      analytics.createEvent(event: event),
+      if (firebaseAnalytics != null)
+        firebaseAnalytics!.createEvent(event: event),
     ]);
   }
 }