+++ lib/src/data/drivers/dio_client_driver.dart
@@ -10,7 +10,7 @@ import '../../domain/interfaces/either.dart';
 import '../../infra/drivers/i_http_driver.dart';
 
 class DioClientDriver extends IHttpDriver with Disposable {
-  DioClientDriver({required this.client,  this.crashLog});
+  DioClientDriver({required this.client, this.crashLog});
 
   final dio.Dio client;
   final CrashLog? crashLog;
@@ -75,8 +75,10 @@ class DioClientDriver extends IHttpDriver with Disposable {
         null => null,
       },
       headers: {
-        if (options != null)
+        if (options != null && options.accessToken != null)
           'Authorization': '${options.accessTokenType} ${options.accessToken}',
+        if (options != null && options.apiKey != null)
+          '${options.apiMapKey}': '${options.apiKey}',
         ...?options?.extraHeaders,
       },
     );