+++ lib/src/domain/entities/pagination_entity.dart
@@ -2,31 +2,31 @@ import '../enums/sorting_order_type_enum.dart';
 
 class PaginationEntity {
   PaginationEntity({
-    this.pageNumber = 1,
-    this.limit = 10,
-    this.offset = 0,
-    this.sortingOrder = SortingOrderType.desc,
+    this.limit,
+    this.offset,
+    this.pageNumber,
+    this.sortingOrder,
     this.sortingBy = '',
   });
-  int pageNumber;
   int? limit;
   int? offset;
-  SortingOrderType sortingOrder;
+  int? pageNumber;
   String sortingBy;
+  SortingOrderType? sortingOrder;
 
   PaginationEntity copyWith({
-    int? pageNumber,
     int? limit,
     int? offset,
-    SortingOrderType? sortingOrder,
+    int? pageNumber,
     String? sortingBy,
+    SortingOrderType? sortingOrder,
   }) {
     return PaginationEntity(
-      pageNumber: pageNumber ?? this.pageNumber,
       limit: limit ?? this.limit,
       offset: offset ?? this.offset,
-      sortingOrder: sortingOrder ?? this.sortingOrder,
       sortingBy: sortingBy ?? this.sortingBy,
+      pageNumber: pageNumber ?? this.pageNumber,
+      sortingOrder: sortingOrder ?? this.sortingOrder,
     );
   }
 }