+++ lib/src/domain/services/i_location_service.dart
@@ -1,4 +1,4 @@
-import '../../../core.dart';
+import '../../../base_core.dart';
 
 abstract class ILocationService {
   Future<Either<ILocationFailure, PositionEntity>> getCurrentPosition();