+++ lib/src/infra/drivers/i_location_driver.dart
@@ -1,4 +1,4 @@
-import '../../../core.dart';
+import '../../../base_core.dart';
 
 abstract class ILocationDriver {
   Future<Either<ILocationFailure, PositionEntity>> getCurrentPosition();