+++ lib/src/infra/drivers/i_share_driver.dart
@@ -4,7 +4,6 @@ import 'package:flutter/widgets.dart';
 
 import '../../domain/interfaces/either.dart';
 
-
 abstract class IShareDriver {
   Future<Either<Exception, Unit>> shareText({required String text});
   Future<Either<Exception, Unit>> shareFiles({required List<File> files});