+++ lib/src/infra/models/received_notifications_model.dart
@@ -13,13 +13,12 @@ class ReceivedNotificationModel extends ReceivedNotificationEntity
     required super.payload,
   });
 
-
   factory ReceivedNotificationModel.fromMap(Map<String, dynamic> map) {
     return ReceivedNotificationModel(
-      id: map['id']?.toInt() ?? 0,
       title: map['title'],
       body: map['body'],
       payload: map['payload'],
+      id: int.tryParse(map['id'].toString()) ?? 0,
     );
   }