+++ CHANGELOG.md
@@ -1,3 +1,20 @@
-## 0.0.1
+## 1.0.4
 
-* TODO: Describe initial release.
+* Fix token authorization and api key into http client
+* Update pagination params
+
+## 1.0.3
+
+* Improve Local auth driver
+
+## 1.0.2
+
+* Update EventLog
+
+## 1.0.1
+
+* Update dependencies and improve documentation
+
+## 1.0.0
+
+* Release production version